move compiler out of pkg/pligos namespace
