+++ pkg/compiler/testdata/a/pligos.yaml
