+++ pkg/compiler/testdata/a/result.yaml
