+++ pkg/compiler/testdata/a/schema.yaml
