+++ pkg/compiler/testdata/b/pligos.yaml
