+++ pkg/compiler/testdata/b/result.yaml
