+++ pkg/compiler/testdata/b/schema.yaml
